Maybe this will fix the flurry crash bug, type is now an integer
git-svn-id: https://trac.beirdo.ca:4443/svn/havokmud/trunk@614 69fa473d-82f3-0310-a870-f1ebe01cfa2b

+++ havokmud/structs.h
@@ -1446,7 +1446,7 @@ struct char_skill_data
 /* Used in CHAR_FILE_U *DO*NOT*CHANGE* */
 struct affected_type
 {
-	short type;           /* The type of spell that caused this      */
+	int type;           /* The type of spell that caused this      */
 	sh_int duration;      /* For how long its effects will last      */
 	long modifier;       /* This is added to apropriate ability     */
 	long location;        /* Tells which ability to change(APPLY_XXX)*/
